+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Features
+
+- Make keyboard shortcut hint in bottom bar clickable to open the command palette
+
 ## [1.41.0]
 
 ### Features